+++ source/renderer/app/themes/daedalus/white.js
@@ -816,6 +816,22 @@ export const WHITE_THEME_OUTPUT = {
     '--theme-transactions-icon-type-pending-warning-background-color':
       'rgba(236, 93, 107, 0.8)',
     '--theme-transactions-state-pending-warning-background-color': '#ec5d6b',
+    '--theme-transactions-filter-button-shadow-color': 'rgba(0, 0, 0, 0.18)',
+    '--theme-transactions-filter-button-color': '#fff',
+    '--theme-transactions-date-picker-button-background-color':
+      'rgba(41, 181, 149, 0.14)',
+    '--theme-transactions-date-picker-button-background-color-hover':
+      'rgba(41, 181, 149, 0.07)',
+    '--theme-transactions-filter-modal-bg-color': '#fff',
+    '--theme-transactions-filter-date-picker-shadow':
+      '0 5px 20px 0 rgba(0, 0, 0, 0.25)',
+    '--theme-transactions-filter-title-button-text-color': '#fafbfc',
+    '--theme-transactions-filter-title-button-background-color':
+      'rgba(94, 96, 102, 0.4)',
+    '--theme-transactions-filter-title-button-background-color-hover':
+      'rgba(94, 96, 102, 0.6)',
+    '--theme-transactions-filter-title-button-background-color-active':
+      'rgba(94, 96, 102, 0.8)',
   },
   uploader: {
     '--theme-uploader-text-color': '#2d2d2d',